Initial Commit
This type/provider is most definitely a work in progress.  As it stands,
it WILL create a plist currently, but it's not idempotent and it's got
lots of flaws (logic and design, I'm sure).  The idea is something like
this:

        property_list_key { 'array_test':
          ensure     => present,
          domain     => '/tmp/com.puppetlabs.puppet',
          key        => 'arraytest',
          value      => ['array', 'values'],
          value_type => 'array',
        }

Right now, the provider has problems determining anything that ISN'T a string - so that needs to be worked out.  Like I mentioned above, it WILL create this plist key/value...but it will create it EVERY puppet run.