Create a Rubycocoa Provider
Previously, I had to wrangle with facter/util/plist when dealing with
plists. Rubycocoa is much better at this task.  This commit adds a
Rubycocoa provider.  It's WORKING, but still not idempotent (as I'm sure
there's problems between the value that Puppet is reading from the
resource and the value that Puppet is getting from the value method).