Add tests for insync?() method
I found that the spec tests didn't accurately represent what
puppet apply would do - specifically around testing whether
a resource was in sync. Because of this, I wrote tests that
will catch that behavior so we won't HAVE to do a puppet
apply run just to test the type (though it's still recommended).